Add more global parameters, new antenna model
The new antennal model takes a balcony railing into account as a big
reflector.
Globally we can now specify the population size, and force the maximum
gain considered during optimization to the horizontal plane and/or the
forward direction and/or the backward direction. If both, forward
direction *and* backward direction are forced, the larger of the two is
used. The latter can be used to prevent the search from finding
configurations where a folded dipole degenerates into a loop (thus no
longer radiating in the direction of the folded dipole but orthogonal to
that direction).