Allow better SWR control during optimization
The max-swr option allows to specify the maximum SWR we want.
New option relax-swr allows to not count the SWR into the evaluation as
long as all three SWR values are below max-swr.
For the folded dipole with balcony reflector we allow a new option that
specifies a larger minimum reflector distance (and allows a larger
reflector distance but this doesn't seem to be used by the optimizer)
Also allow specifying the radii of some of the parts of the balcony: The
elements of the balcony are mostly rectangular not round.
Update statistics with new statistics tool.